Прототип UI. Изменил структура проекта Создал приложение через Tkinter. Создал первую вкладку. Разложил по директориям классы.
