--- module_functions.py
@@ -1,122 +1,0 @@
-from datetime import *
-from dateutil.relativedelta import *
-import calendar
-from module_structures import *
-
-
-def getPeriod(period=' ', number = 0):
-    first_value = datetime.now().date()
-    number -= 1
-    
-    if period == "weeks":
-        first_value = first_value + relativedelta(weeks=(-number), weekday=MO(-1))
-        last_value = first_value + relativedelta(weekday=SU(+1))
-
-    if period == "months":
-        first_value = first_value + relativedelta(months=(-number))
-        first_value += relativedelta(day=1)
-        last_value = first_value + relativedelta(day=31)
-
-    if period == "quarts":        
-        d = 3*number
-        first_value = first_value + relativedelta(day=1, months=-(d))
-        num_m = checkquartint(first_value.month)        
-        first_value = first_value + relativedelta(month=num_m[0])
-        last_value = first_value + relativedelta(day=31, month=num_m[1])
-
-    if period == "years":
-        first_value = first_value + relativedelta(month = 1, day=1, years=-number)
-        last_value = first_value + relativedelta(month = 1, day=1, years=+1)
-    return (first_value, last_value)
-
-def checkquartint(month):
-        quart = ((month-1)//3 + 1)
-        if quart == 1:
-            return [1,3]
-        if quart == 2:
-            return  [4, 6]
-        if quart == 3:
-            return [7, 9]
-        if quart == 4:
-            return [10, 12]
-
-def checkquartNum(month):
-        return ((month-1)//3 + 1)
-        
-
-def finddates(start_t, end_t):
-    value_list = []
-    while True:
-        value_list.append((start_t).strftime("%d/%m/%Y"))
-        start_t = start_t + relativedelta(days=(+1))        
-        if (start_t == end_t):
-            value_list.append((start_t).strftime("%d/%m/%Y"))
-            break
-    return value_list
-
-def findboxvalues(period):
-    names_arr = []
-    for i in range(1,5):
-        start_and_last = getPeriod(period, i)
-        if period == 'weeks':
-                transformed_info1 = start_and_last[0].strftime("%d/%m/%Y")
-                transformed_info2 = start_and_last[1].strftime("%d/%m/%Y")    
-                transformed_info1 = transformed_info1[:6] +  transformed_info1[8:10]
-                transformed_info2 = transformed_info2[:6] +  transformed_info2[8:10]
-                title = '{} - {}'.format(transformed_info1, transformed_info2)
-                names_arr.append(title)
-        if period == 'months':
-                month = nameOfMonth(start_and_last[0].month)                
-                title = '{}  {}'.format(start_and_last[0].year, month)
-                names_arr.append(title)
-        if period == 'quarts':
-                quart = nameOfQuart(checkquartNum(start_and_last[0].month))
-                title = '{}  {}'.format(quart, start_and_last[0].year)
-                names_arr.append(title)
-        if period == 'years':
-                title = '{}'.format(start_and_last[0].year)
-                names_arr.append(title)
-    return names_arr
-
-
-def shorttitle(period, arr_dt):
-    new_l = []
-    if period == 'weeks':
-        for i in range(len(arr_dt)):
-            if i%2 != 0:
-                new_l.append((arr_dt[i])[:-5])
-            else:
-                new_l.append("")   
-    if period == 'months':
-        for i in range(len(arr_dt)): 
-                if int(arr_dt[i][:2])%5 == 0:
-                    new_l.append((arr_dt[i])[:-5])
-                else:
-                    new_l.append("")
-    if period == 'quarts':
-        for i in range(len(arr_dt)):
-            interval = str(arr_dt[i][:2])
-            if (interval == '15' or interval == '01'):
-                if(interval == '15'):
-                    month = nameOfMonth(int(arr_dt[i][3:-5]))
-                    new_l.append(month)
-                else:
-                    new_l.append("")
-        new_l.append("")
-    if period == 'years':
-        for i in range(len(arr_dt)):
-            interval = str(arr_dt[i][0:2])
-            if (interval == '01' or interval == '15'):
-                if(interval == '01'):
-                    elem = (arr_dt[i][3:5])
-                    elem = int(elem)
-                    month = (nameOfMonth(elem))[0:3]
-                    new_l.append(month)
-                else:
-                    new_l.append("")
-    return new_l
-    
-            
-
-# arr = ['01/01/2021', '02/01/2021', '03/01/2021', '04/01/2021', '05/01/2021', '06/01/2021', '07/01/2021', '08/01/2021', '09/01/2021', '10/01/2021', '11/01/2021', '12/01/2021', '13/01/2021', '14/01/2021', '15/01/2021', '16/01/2021', '17/01/2021', '18/01/2021', '19/01/2021', '20/01/2021', '21/01/2021', '22/01/2021', '23/01/2021', '24/01/2021', '25/01/2021', '26/01/2021', '27/01/2021', '28/01/2021', '29/01/2021', '30/01/2021', '31/01/2021', '01/02/2021', '02/02/2021', '03/02/2021', '04/02/2021', '05/02/2021', '06/02/2021', '07/02/2021', '08/02/2021', '09/02/2021', '10/02/2021', '11/02/2021', '12/02/2021', '13/02/2021', '14/02/2021', '15/02/2021', '16/02/2021', '17/02/2021', '18/02/2021', '19/02/2021', '20/02/2021', '21/02/2021', '22/02/2021', '23/02/2021', '24/02/2021', '25/02/2021', '26/02/2021', '27/02/2021', '28/02/2021', '01/03/2021', '02/03/2021', '03/03/2021', '04/03/2021', '05/03/2021', '06/03/2021', '07/03/2021', '08/03/2021', '09/03/2021', '10/03/2021', '11/03/2021', '12/03/2021', '13/03/2021', '14/03/2021', '15/03/2021', '16/03/2021', '17/03/2021', '18/03/2021', '19/03/2021', '20/03/2021', '21/03/2021', '22/03/2021', '23/03/2021', '24/03/2021', '25/03/2021', '26/03/2021', '27/03/2021', '28/03/2021', '29/03/2021', '30/03/2021', '31/03/2021', '01/04/2021', '02/04/2021', '03/04/2021', '04/04/2021', '05/04/2021', '06/04/2021', '07/04/2021', '08/04/2021', '09/04/2021', '10/04/2021', '11/04/2021', '12/04/2021', '13/04/2021', '14/04/2021', '15/04/2021', '16/04/2021', '17/04/2021', '18/04/2021', '19/04/2021', '20/04/2021', '21/04/2021', '22/04/2021', '23/04/2021', '24/04/2021', '25/04/2021', '26/04/2021', '27/04/2021', '28/04/2021', '29/04/2021', '30/04/2021', '01/05/2021', '02/05/2021', '03/05/2021', '04/05/2021', '05/05/2021', '06/05/2021', '07/05/2021', '08/05/2021', '09/05/2021', '10/05/2021', '11/05/2021', '12/05/2021', '13/05/2021', '14/05/2021', '15/05/2021', '16/05/2021', '17/05/2021', '18/05/2021', '19/05/2021', '20/05/2021', '21/05/2021', '22/05/2021', '23/05/2021', '24/05/2021', '25/05/2021', '26/05/2021', '27/05/2021', '28/05/2021', '29/05/2021', '30/05/2021', '31/05/2021', '01/06/2021', '02/06/2021', '03/06/2021', '04/06/2021', '05/06/2021', '06/06/2021', '07/06/2021', '08/06/2021', '09/06/2021', '10/06/2021', '11/06/2021', '12/06/2021', '13/06/2021', '14/06/2021', '15/06/2021', '16/06/2021', '17/06/2021', '18/06/2021', '19/06/2021', '20/06/2021', '21/06/2021', '22/06/2021', '23/06/2021', '24/06/2021', '25/06/2021', '26/06/2021', '27/06/2021', '28/06/2021', '29/06/2021', '30/06/2021', '01/07/2021', '02/07/2021', '03/07/2021', '04/07/2021', '05/07/2021', '06/07/2021', '07/07/2021', '08/07/2021', '09/07/2021', '10/07/2021', '11/07/2021', '12/07/2021', '13/07/2021', '14/07/2021', '15/07/2021', '16/07/2021', '17/07/2021', '18/07/2021', '19/07/2021', '20/07/2021', '21/07/2021', '22/07/2021', '23/07/2021', '24/07/2021', '25/07/2021', '26/07/2021', '27/07/2021', '28/07/2021', '29/07/2021', '30/07/2021', '31/07/2021', '01/08/2021', '02/08/2021', '03/08/2021', '04/08/2021', '05/08/2021', '06/08/2021', '07/08/2021', '08/08/2021', '09/08/2021', '10/08/2021', '11/08/2021', '12/08/2021', '13/08/2021', '14/08/2021', '15/08/2021', '16/08/2021', '17/08/2021', '18/08/2021', '19/08/2021', '20/08/2021', '21/08/2021', '22/08/2021', '23/08/2021', '24/08/2021', '25/08/2021', '26/08/2021', '27/08/2021', '28/08/2021', '29/08/2021', '30/08/2021', '31/08/2021', '01/09/2021', '02/09/2021', '03/09/2021', '04/09/2021', '05/09/2021', '06/09/2021', '07/09/2021', '08/09/2021', '09/09/2021', '10/09/2021', '11/09/2021', '12/09/2021', '13/09/2021', '14/09/2021', '15/09/2021', '16/09/2021', '17/09/2021', '18/09/2021', '19/09/2021', '20/09/2021', '21/09/2021', '22/09/2021', '23/09/2021', '24/09/2021', '25/09/2021', '26/09/2021', '27/09/2021', '28/09/2021', '29/09/2021', '30/09/2021', '01/10/2021', '02/10/2021', '03/10/2021', '04/10/2021', '05/10/2021', '06/10/2021', '07/10/2021', '08/10/2021', '09/10/2021', '10/10/2021', '11/10/2021', '12/10/2021', '13/10/2021', '14/10/2021', '15/10/2021', '16/10/2021', '17/10/2021', '18/10/2021', '19/10/2021', '20/10/2021', '21/10/2021', '22/10/2021', '23/10/2021', '24/10/2021', '25/10/2021', '26/10/2021', '27/10/2021', '28/10/2021', '29/10/2021', '30/10/2021', '31/10/2021', '01/11/2021', '02/11/2021', '03/11/2021', '04/11/2021', '05/11/2021', '06/11/2021', '07/11/2021', '08/11/2021', '09/11/2021', '10/11/2021', '11/11/2021', '12/11/2021', '13/11/2021', '14/11/2021', '15/11/2021', '16/11/2021', '17/11/2021', '18/11/2021', '19/11/2021', '20/11/2021', '21/11/2021', '22/11/2021', '23/11/2021', '24/11/2021', '25/11/2021', '26/11/2021', '27/11/2021', '28/11/2021', '29/11/2021', '30/11/2021', '01/12/2021', '02/12/2021', '03/12/2021', '04/12/2021', '05/12/2021', '06/12/2021', '07/12/2021', '08/12/2021', '09/12/2021', '10/12/2021', '11/12/2021', '12/12/2021', '13/12/2021', '14/12/2021', '15/12/2021', '16/12/2021', '17/12/2021', '18/12/2021', '19/12/2021', '20/12/2021', '21/12/2021', '22/12/2021', '23/12/2021', '24/12/2021', '25/12/2021', '26/12/2021', '27/12/2021', '28/12/2021', '29/12/2021', '30/12/2021', '31/12/2021', '01/01/2022']
-# print(shorttitle("years", arr))
